refactor(LocalAddressBook): move header handlers into SectionHeader
