--- src/pages/options/AddressesSettings/DeleteSectionDialog.tsx
@@ -1,41 +1,0 @@
-import {
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogContentText,
-  DialogTitle,
-  Button,
-} from '@mui/material';
-
-interface DeleteSectionDialogProps {
-  open: boolean;
-  onClose: () => void;
-  onConfirm: () => void;
-}
-
-export default function DeleteSectionDialog({
-  open,
-  onClose,
-  onConfirm,
-}: DeleteSectionDialogProps) {
-  return (
-    <Dialog
-      open={open}
-      onClose={onClose}
-      aria-labelledby="delete-dialog-title"
-      aria-describedby="delete-dialog-description">
-      <DialogTitle id="delete-dialog-title">Delete Section?</DialogTitle>
-      <DialogContent>
-        <DialogContentText id="delete-dialog-description">
-          Are you sure you want to delete this section? This action cannot be undone.
-        </DialogContentText>
-      </DialogContent>
-      <DialogActions>
-        <Button onClick={onClose}>Cancel</Button>
-        <Button onClick={onConfirm} color="error" variant="contained">
-          Delete
-        </Button>
-      </DialogActions>
-    </Dialog>
-  );
-}
